+++ src/java.base/share/classes/jdk/internal/util/OperatingSystemProps.java.template
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.internal.misc;
+package jdk.internal.util;
 
 /**
  * The corresponding source file is generated by GensrcMisc.gmk for java.base.
@@ -36,7 +36,7 @@ class OperatingSystemProps {
     private static final int TARGET_OS_aix     = 3;
 
     // Index/ordinal of the current OperatingSystem enum as substituted by the build
-    static final int TARGET_OS_ORDINAL = TARGET_OS_@@OPENJDK_TARGET_OS@@;
+    static final int CURRENT_OS_ORDINAL = TARGET_OS_@@OPENJDK_TARGET_OS@@;
 
     // Precomputed booleans for each Operating System
     static final boolean TARGET_OS_IS_LINUX   = TARGET_OS_@@OPENJDK_TARGET_OS@@ == TARGET_OS_linux;
